use CSS.registerProperty for unit convertions
