+++ packages/container-query/src/index.js
@@ -1,9 +1,25 @@
 import { useState, useEffect, useRef } from 'react'
 import { useThemeUI } from 'theme-ui'
-import { getBreakpointIndex, unit2px } from './utils'
 
 // Shared with @styled-system/css
 const defaultBreakpoints = [40, 52, 64].map(n => n + 'em')
+
+export const getBreakpointIndex = (ar = [], v) => {
+  const len = ar.length
+  for (let i = 0; i < len; i++) {
+    if (v <= ar[i]) {
+      return i
+    }
+  }
+  return len
+}
+
+CSS.registerProperty({
+  name: '--bk-size',
+  syntax: '<length>',
+  inherits: false,
+  initialValue: 0,
+})
 
 export const useContainerQuery = componentBreakpoints => {
   const { theme = {} } = useThemeUI()
@@ -21,7 +37,10 @@ export const useContainerQuery = componentBreakpoints => {
       return
     }
 
-    const normalizedBreakpoints = breakpoints.map(bk => unit2px(el, bk))
+    const normalizedBreakpoints = breakpoints.map(bk => {
+      el.style.setProperty('--bk-size', bk)
+      return el.computedStyleMap().get('--bk-size').value
+    })
 
     const handler = entries => {
       const entry = entries[0]
