+++ packages/container-query/test/index.js
@@ -1,6 +1,6 @@
 /** @jsx jsx */
 import { render, cleanup, act } from '@testing-library/react'
-import { useContainerQuery } from '../src'
+import { useContainerQuery, getBreakpointIndex } from '../src'
 
 const { jsx } = jest.requireActual('theme-ui')
 
@@ -182,5 +182,35 @@ describe('useContainerQuery', () => {
     })
 
     expect(tree.getByText('red')).toBeTruthy()
+  })
+})
+
+describe('getBreakpointIndex', () => {
+  test('width is less than the min breakpoint', () => {
+    expect(getBreakpointIndex([480, 640, 1024], 320)).toEqual(0)
+  })
+
+  test('width is equal to the min breakpoint', () => {
+    expect(getBreakpointIndex([480, 640, 1024], 480)).toEqual(0)
+  })
+
+  test('width is in the range, on the left side', () => {
+    expect(getBreakpointIndex([480, 640, 1024], 520)).toEqual(1)
+  })
+
+  test('width is bigger than the max breakpoint', () => {
+    expect(getBreakpointIndex([480, 640, 1024], 1280)).toEqual(3)
+  })
+
+  test('width is equal to the max breakpoint', () => {
+    expect(getBreakpointIndex([480, 640, 1024], 1024)).toEqual(2)
+  })
+
+  test('width is in the range, on the right side', () => {
+    expect(getBreakpointIndex([480, 640, 1024], 960)).toEqual(2)
+  })
+
+  test('array is empty', () => {
+    expect(getBreakpointIndex([], 960)).toEqual(0)
   })
 })
